update
    - refactored navbar and added notifications' menu component
    - notifications menu should now fetch only unread messages
